Extract CovenSelectCard to its own file, byte-for-byte (#2325)
The pure move, so the repaint that follows has nothing structural in it.
`COVENSelectCard` leaves `FactionSelectCard.tsx` for
`components/selectCard/CovenSelectCard.tsx`, matching how
`components/taskCard/` is laid out and following the shape #2322 set for
S.N.I.D.E. The function is renamed to `CovenSelectCard` so the file, the
export and the manifest row all read the way `CovenTaskCard` does; the
manifest row keeps its shape and drops the
`.then((m) => ({ default: m.COVENSelectCard }))` unwrap, because the new
file default-exports it.

PROOF THE DIRECTORY IS UNCHANGED. `renderToStaticMarkup` of
`<FactionSelectCard faction="coven" />` across all six state x members
combinations is byte-identical before and after this commit (19,149 bytes,
`diff` clean). The probe was a throwaway; the standing guard is
`factionSelectCardDispatch.test.tsx`, green.

The dispatcher's `import * as coven from "../factionMarks/covenSlip"` goes
with the block — no other tile read it. Nothing else in the file moved:
`CAST`, the shared types, `LEGACY_SLUG` and the seven other archetypes are
untouched, so the siblings extracting beside this one (#2323/#2324/#2326)
land on non-adjacent hunks.
