+++ frontend/src/factions/coven.ts
@@ -29,12 +29,12 @@ const CovenPraxisCard = lazyArchetype(() => import('../components/praxisCard/des
 const CovenScoreStamp = lazyArchetype(() => import('../components/praxisCard/scoreStamp/CovenScoreStamp'))
 const CovenSeal = lazyArchetype(() => import('../components/metataskSeal/skins/CovenSeal'))
 const CovenSigil = lazyArchetype(() => import('../components/sigil/CovenSigil').then((m) => ({ default: m.CovenSigil })))
-const COVENSelectCard = lazyArchetype(() => import('../components/selectCard/FactionSelectCard').then((m) => ({ default: m.COVENSelectCard })))
+const CovenSelectCard = lazyArchetype(() => import('../components/selectCard/CovenSelectCard'))
 
 export const COVEN_MANIFEST: FactionManifest = {
   slug: 'coven',
 
-  factionSelectCard: () => COVENSelectCard,
+  factionSelectCard: () => CovenSelectCard,
   taskCard: () => CovenTaskCard,
   praxisCard: () => CovenPraxisCard,
   scoreStamp: () => CovenScoreStamp,
